PCIEXBAR: Removing hardcoding address for each PCI host
Reduce the low memory size to 1G so that we reserve 3G for PCI.
For multiple PCI segment, the first MCFG base is from 0x40000000
and dynamically allocates the MCFG space for each segment.
The PCI hole is just behind all the MCFG space according to the
segment number.

This corresponds to the same logic in OVMF changes.

Note:
1. We can't enable MCFG above 4G, so only use below 4G
2. If each MCFG use 256M, we could only set 4 PCI segment.

Signed-off-by: Jing Liu <jing2.liu@linux.intel.com>